Removed catalog object lookups (getObject) from index_html for ATVideo objects Only use the brain object!
Made the template compatiable with existing, except for formatted duration , since thats a call on an object
The existing check on thumbnail image is maintained via a new index on the catalog, which makes it avail on the brain object
returned by catalog searches.

Inside ATVideo content object, i made hasThumbnailImage an catalog index, via a schema change on ATVideo

(Reinstall, and do a 'clear and rebuild' on existing catalogs via ZMI)

minor change to the external file system , file paths, fixup script.