Added support for multiple syntax modes with a default mode pointing to tal-like syntax, subclassed the py-tal-like parser to accept the tal namespace instead of the py namespace.
Added configurable support for hacking, development and production mode, where the later two are determined by the Zope development mode. In hacking mode the template byte code is always regenerated so changes to the parser logic are taken into account, in development mode it is only regenerated if the template itself changed and in production mode it is never regenerated but the potentially stale byte code file is taken.

Also added a default optimization level and made in configurable per template class. Currently it defaults to O1 as O2 and O3 produce some errors for one of the templates.