Install video_listing_view as option in display menu Fix for http://www.plumi.org/ticket/117 - used http://plone.org/products/diyplonestyle/issues/8 for guidance
fixed the bug where the new five view is being used, and navigation macros not working properly
http://trac.engagemedia.org/projects/ticket/503
http://www.nabble.com/template_id-vs-action-id-with-Five-views-to3973672.html
quote: 
"
When I've neded, this I explicitly defined a template_id in my
template, which is lame but works.  Plone should really be looking at
the url instead of template_id, and will do so in 2.5 (hopefully). 
"
so ive hardcoded "template_id" in template :((