Correct type registration.
A very common pattern when registering content class for Archetypes based
content types is code like this:

{{{
for atype, constructor in allTypes:
    kind = "%s: %s" % (PROJECTNAME, atype.archetype_name)
    ContentInit(
        kind,
}}}

This relies heavily on atype.archetype_name being unique for every type
in your product. Now that we use !GenericSetup to setup content types
there is no reason for a class to have an archetypes_name class variable
anymore, except for this hidden secret. Luckily we can substitute
atype.archetype_name with atype.__name__ which returns the name of the
class. This is also much more in line with Zope2 standards which expect
to have something related to the meta type here, not related to some
descriptive string for humans such as archetype_name.

The corrected code, which works in both pre-GenericSetup and !GenericSetup
based products is:

{{{
for atype, constructor in allTypes:
    kind = "%s: %s" % (PROJECTNAME, atype.__name__)
    ContentInit(
        kind,
}}}