fix white space in lines related to `checkCreationFlag` so i can comment on some investigations made:
after implementing a check for the "creation flag" (in `filterTemporaryItems`), which was supposed to prevent not-yet-fully-created objects from unnecessarily being added to the queue, it turns out that the first time an add event is fired for a new object the flag isn't set yet at all (in `ObjectManager._setObject`).  so the object always ends up being queued, even with filtering on the flag.  given that, the extending the filter doesn't really make sense, of course, since it would essentially just save a single reduce operation anyway...