If the absence starts tomorrow, pretend the employee is already sick for a day.
Used the .toordinal method to compare the start and end date since I'm
not sure of the type of the start date. It can be an offset-naive or
offset-aware datetime or it can be a date. To prevent problems I've
decided to just compare the "proleptic Gregorian ordinal" to compare
the dates.